tests/production-grade-api-test/requirements.txt
pytest==8.2.2
pyjwt==2.9.0
pytest-dotenv
pydantic==2.8.2
requests